+++ src/main/scala/com/programmaticallyspeaking/ncd/nashorn/NashornDebuggerHost.scala
@@ -264,12 +264,20 @@ class NashornDebuggerHost(val virtualMachine: XVirtualMachine, protected val asy
   }
 
   private def cleanupPausing(): Unit = {
-//    pausedData.foreach()
+    // We evaluate code with all ClassPrepare requests disabled to avoid deadlock. However, code evaluation mau
+    // result in an added script (e.g. when the 'load' extension is used). Go through added classes and let the
+    // scanner decide if scripts were added.
+    val addedClasses = pausedData.toSeq.flatMap(_.addedClasses())
     pausedData = None
     mappingRegistry.clear() // only valid when paused
 
     // Enable exception requests again (see prepareForPausing)
     virtualMachine.eventRequestManager().exceptionRequests().asScala.foreach(_.enable())
+
+    if (addedClasses.nonEmpty) {
+      log.debug(s"Scanning ${addedClasses.size} classes to detect scripts added during code evaluation.")
+      addedClasses.foreach(_scanner.consider)
+    }
   }
 
   private def maybeEmitErrorEvent(pd: PausedData): Unit = pd.exceptionEventInfo match {