+++ src/main/scala/com/programmaticallyspeaking/ncd/nashorn/PausedData.scala
@@ -3,12 +3,22 @@ package com.programmaticallyspeaking.ncd.nashorn
 import com.programmaticallyspeaking.ncd.host.types.ObjectPropertyDescriptor
 import com.programmaticallyspeaking.ncd.host.{ObjectId, StackFrame}
 import com.programmaticallyspeaking.ncd.nashorn.NashornDebuggerHost.ObjectPropertiesKey
-import com.sun.jdi.ThreadReference
+import com.sun.jdi.{ReferenceType, ThreadReference}
 import com.sun.jdi.event.{ExceptionEvent, LocatableEvent}
 
 import scala.collection.mutable
 
 private[nashorn] class PausedData(val thread: ThreadReference, val marshaller: Marshaller, stackBuilder: StackBuilder, event: LocatableEvent) {
+  import scala.collection.JavaConverters._
+
+  // Capture known classes when pausing, to be able to diff afterwards. This is currently the only way I know to
+  // detect classes added during code evaluation.
+  private val classesBefore = thread.virtualMachine().allClasses().asScala.toSet
+
+  def addedClasses(): Seq[ReferenceType] = {
+    val classesNow = thread.virtualMachine().allClasses().asScala.toSet
+    classesNow.diff(classesBefore).toSeq
+  }
 
   /** We assume that we can cache object properties as long as we're in a paused state. Since we're connected to a
     * Java process, an arbitrary Java object may change while in this state, so we only cache JS objects.