+++ src/test/scala/com/programmaticallyspeaking/ncd/nashorn/EvaluateTestFixture.scala
@@ -11,12 +11,12 @@ class EvaluateTestFixture extends UnitTest with NashornScriptHostTestFixture {
 
   override implicit val executionContext: ExecutionContext = ExecutionContext.global
 
-  protected def evaluateInScript(script: String)(tester: (ScriptHost, Seq[StackFrame]) => Unit): Unit = {
+  protected def evaluateInScript(script: String, unknownEventHandler: (ScriptEvent) => Unit = _ => {})(tester: (ScriptHost, Seq[StackFrame]) => Unit): Unit = {
     assert(script.contains("debugger;"), "Script must contain a 'debugger' statement")
     val stackframesPromise = Promise[Seq[StackFrame]]()
     val observer = Observer.from[ScriptEvent] {
       case bp: HitBreakpoint => stackframesPromise.trySuccess(bp.stackFrames)
-      case _ => // ignore
+      case x => unknownEventHandler(x)
     }
     // Newline after $script so that comments won't cause syntax error
     val wrapper =