+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Don't rely on Nashorn syntax extensions in the property extractor code, since the target may
   have disabled them (issue #68)
 * Pause on reference errors if pausing on exceptions.
+* Detect a script added during code evaluation, e.g. when the `load` extension is used (issue #66)
 
 ## 0.6.0 (2017-10-22)